fix - error on history model when install this project
