--- machado/history/models.py
@@ -1,11 +1,0 @@
-from django.db import models
-from django.contrib.auth.models import User
-
-class History(models.Model):
-    user = models.ForeignKey(User, on_delete=models.DO_NOTHING)
-    description = models.CharField(max_length=255)
-    method = models.CharField(max_length=50)
-    created_at = models.DateTimeField(auto_now_add=True)
-
-    def __str__(self):
-        return self.action
